+++ components/collection/drop/GenerativePreview.vue
@@ -1,64 +1,122 @@
 <template>
-  <div class="fixed-size border">
-    <div
-      class="fixed-top-left border px-4 py-2 bg-background-color z-[1] whitespace-nowrap">
-      {{ $t('mint.unlockable.yourVariation') }}
-    </div>
-
+  <div
+    :class="[
+      'border bg-background-color shadow-primary p-5 pb-6',
+      'w-full h-min',
+      'lg:max-w-[490px]',
+    ]">
     <BaseMediaItem
       :src="sanitizeIpfsUrl(displayUrl)"
       :mime-type="generativeImageUrl ? 'text/html' : ''"
       preview
       is-detail
-      class="border-b" />
-    <div class="flex justify-center items-center py-6">
-      <NeoButton
-        v-if="isLoading"
-        class="border-k-grey hover-button fixed-width pointer-events-auto cursor-wait"
-        rounded
-        no-shadow
-        disabled>
-        {{ $t('mint.unlockable.generating') }}
-        <NeoIcon icon="circle-notch" spin />
-      </NeoButton>
-      <NeoButton
-        v-else
-        class="border-k-grey hover-button fixed-width"
-        rounded
-        no-shadow
-        icon="arrow-rotate-left"
-        @click="generateNft()">
-        {{ $t('mint.unlockable.variations') }}
-      </NeoButton>
+      class="border" />
 
-      <a
-        v-safe-href="sanitizeIpfsUrl(displayUrl)"
-        class="flex items-center text-k-blue hover:text-k-blue-hover fixed-right"
-        rel="nofollow noopener noreferrer"
-        target="_blank"
-        role="link">
-        {{ $t('img') }}
-        <NeoIcon icon="arrow-up-right" class="ml-1" />
-      </a>
+    <NeoButton
+      v-if="isLoading"
+      class="mt-5 h-[40px] border-k-grey pointer-events-auto cursor-wait hover:!bg-transparent"
+      expanded
+      rounded
+      no-shadow
+      disabled>
+      <div class="inline-flex items-center">
+        <span class="mr-2">{{ $t('mint.unlockable.generating') }}</span>
+        <NeoIcon icon="circle-notch" spin />
+      </div>
+    </NeoButton>
+
+    <NeoButton
+      v-else
+      class="mt-5 h-[40px] border-k-grey hover:!bg-transparent"
+      expanded
+      rounded
+      no-shadow
+      @click="generateNft()">
+      <div class="inline-flex items-center">
+        <span>{{ $t('drops.createNewVariation') }}</span>
+        <NeoIcon icon="arrow-rotate-left" class="ml-2" />
+      </div>
+    </NeoButton>
+
+    <hr class="my-5" />
+
+    <div class="flex justify-between items-center mb-4">
+      <div class="font-bold">
+        <span v-if="!!Number(drop.price)">{{ formattedPrice }}</span>
+        <span v-else>{{ $t('free') }}</span>
+      </div>
+      <div class="flex justify-end items-center">
+        <div class="mr-4 text-neutral-7">{{ mintedPercent }}% ~</div>
+        <div class="font-bold">
+          {{ mintedCount }}/{{ maxCount }}
+          {{ $t('statsOverview.minted') }}
+        </div>
+      </div>
     </div>
+
+    <CollectionUnlockableSlider
+      class="text-neutral-5"
+      :value="mintedCount / maxCount" />
+
+    <CollectionDropMintButton
+      class="mt-6"
+      :collection-id="collectionId"
+      :user-minted-nft-id="userMintedNftId"
+      :is-wallet-connecting="isWalletConnecting"
+      :is-image-fetching="isImageFetching"
+      :is-loading="isLoading"
+      :minimum-funds="minimumFunds"
+      :max-count="maxCount"
+      :mint-count-available="mintCountAvailable"
+      :mint-button="mintButton"
+      :holder-of-collection="holderOfCollection"
+      @mint="emit('mint')" />
   </div>
 </template>
 
 <script setup lang="ts">
+import { blake2AsHex, encodeAddress } from '@polkadot/util-crypto'
 import { NeoButton, NeoIcon } from '@kodadot1/brick'
 import { sanitizeIpfsUrl } from '@/utils/ipfs'
+import { DropItem } from '@/params/types'
+import type {
+  HolderOfCollectionProp,
+  MinimumFundsProp,
+  MintButtonProp,
+} from '@/components/collection/drop/types'
 import { getRandomIntFromRange } from '../unlockable/utils'
-import { blake2AsHex, encodeAddress } from '@polkadot/util-crypto'
 
 const props = defineProps<{
+  drop: DropItem
   minted: number
-  content: string
-  image?: string
+  collectionId: string
+  mintedCount: number
+  mintCountAvailable: boolean
+  maxCount: number
+  minimumFunds: MinimumFundsProp
+  isImageFetching: boolean
+  isWalletConnecting: boolean
+  isLoading: boolean
+  mintButton: MintButtonProp
+  userMintedNftId?: string
+  holderOfCollection?: HolderOfCollectionProp
 }>()
 
-const emit = defineEmits(['select'])
+const emit = defineEmits(['select', 'mint'])
 
 const { accountId } = useAuth()
+const { chainSymbol, decimals } = useChain()
+
+const mintedPercent = computed(() => {
+  const percent = (props.mintedCount / props.maxCount) * 100
+  return Math.round(percent)
+})
+
+const { formatted: formattedPrice } = useAmount(
+  computed(() => props.drop.price),
+  decimals,
+  chainSymbol,
+)
 
 const STEP = 64
 const entropyRange = computed<[number, number]>(() => [
@@ -79,17 +137,17 @@ const getHash = (isDefault?: boolean) => {
 }
 
 const generativeImageUrl = ref(
-  accountId.value ? `${props.content}/?hash=${getHash(true)}` : '',
+  accountId.value ? `${props.drop.content}/?hash=${getHash(true)}` : '',
 )
 
 const isLoading = ref(false)
 
 const displayUrl = computed(() => {
-  return generativeImageUrl.value || props.image
+  return generativeImageUrl.value || props.drop.image
 })
 const generateNft = (isDefault: boolean = false) => {
   isLoading.value = true
-  const metadata = `${props.content}/?hash=${getHash(isDefault)}`
+  const metadata = `${props.drop.content}/?hash=${getHash(isDefault)}`
   generativeImageUrl.value = metadata
   emit('select', generativeImageUrl.value)
 
@@ -108,50 +166,3 @@ watch(
   },
 )
 </script>
-
-<style scoped lang="scss">
-@import '@/assets/styles/abstracts/variables';
-
-.fixed-size {
-  width: 36rem;
-  height: min-content;
-  position: relative;
-
-  @include mobile {
-    width: 100% !important;
-    height: 100% !important;
-    max-width: 35rem;
-  }
-  @include tablet-only {
-    width: 48rem;
-  }
-}
-
-.fixed-top-left {
-  @apply absolute rounded-[3rem] left-[26px] -top-3.5;
-
-  @include mobile {
-    left: 50%;
-    transform: translateX(-50%);
-  }
-}
-
-.hover-button:hover {
-  background: unset;
-  @include ktheme() {
-    border-color: theme('border-color') !important;
-  }
-}
-
-.fixed-right {
-  position: absolute;
-  right: 2rem;
-  @include mobile {
-    right: 1rem;
-  }
-}
-
-.fixed-width {
-  width: 10rem;
-}
-</style>
