+++ components/collection/drop/requirement/Item.vue
