+++ components/collection/unlockable/UnlockableCollectionInfo.vue
@@ -11,7 +11,7 @@
       <div class="flex justify-between items-center">
         <NeoButton
           v-if="hasSeeAllDescriptionOption"
-          class="no-shadow is-text text-left p-0 underline"
+          class="no-shadow is-text text-left p-0"
           :label="seeAllDescription ? $t('showLess') : $t('showMore')"
           @click="toggleSeeAllDescription" />
         <NeoButton
