+++ composables/useViewport.ts
@@ -1,7 +1,10 @@
 import { useWindowSize } from '@vueuse/core'
 
 export default function () {
-  const isMobile = computed(() => useWindowSize().width.value <= 768)
+  const { width } = useWindowSize()
+  const isMobile = computed(() => width.value <= 768)
+  const isDesktop = computed(() => width.value >= 1024)
+  const isFullHD = computed(() => width.value >= 1440)
 
-  return { isMobile }
+  return { isMobile, isDesktop, isFullHD }
 }
